Remove scabbard client's `submit` module
Removes the `splinter::service::scabbard::client::submit` module and
moves the code the main `client` module. The `submit_batches` method is
inlined because it is only used once and it is small.
